+++ tests/test_seirs.py
@@ -34,15 +34,17 @@ def run_model(interactive: bool = False, params: dict | None = None) -> Model:
     seeds = np.maximum(np.minimum(25, scenario.S.values), (scenario.S.values * 0.01).astype(int))
     scenario["S"] -= seeds
     scenario["I"] += seeds
-    p = PropertySet({
-        "nticks": 5 * 365,
-        # "beta": 1.386/7.0, # 1.386 new infections per existing infection every 7 ticks
-        "beta": 1.5 / 7.0,
-        "r_progression": 1.0 / 7.0,  # 7 ticks of incubation (exposure)
-        "r_recovery": 1.0 / 7.0,  # 7 ticks to recovery
-        "r_waning": 1.0 / 182.5,  # 1/2 year to waning
-        **(params or {}),
-    })
+    p = PropertySet(
+        {
+            "nticks": 5 * 365,
+            # "beta": 1.386/7.0, # 1.386 new infections per existing infection every 7 ticks
+            "beta": 1.5 / 7.0,
+            "r_progression": 1.0 / 7.0,  # 7 ticks of incubation (exposure)
+            "r_recovery": 1.0 / 7.0,  # 7 ticks to recovery
+            "r_waning": 1.0 / 182.5,  # 1/2 year to waning
+            **(params or {}),
+        }
+    )
     model = Model(scenario, p)
 
     betas = ValuesMap.from_scalar(p.beta, p.nticks, len(scenario))
@@ -101,6 +103,7 @@ def test_seirs() -> None:
 
 
 if __name__ == "__main__":
+
     def _parse_value(s: str) -> int | float:
         try:
             return int(s)