+++ tests/test_sirs.py
@@ -34,13 +34,15 @@ def run_model(interactive: bool = False, params: dict | None = None) -> Model:
     seeds = np.maximum(np.minimum(25, scenario.S.values), (scenario.S.values * 0.01).astype(int))
     scenario["S"] -= seeds
     scenario["I"] += seeds
-    p = PropertySet({
-        "nticks": 5 * 365,
-        "beta": 1.5 / 7.0,  # 1.25 new infections per existing infection every 7 ticks
-        "r_recovery": 1.0 / 7.0,  # 7 days to recovery
-        "r_waning": 1.0 / 30.0,  # 30 days to susceptibility
-        **(params or {}),
-    })
+    p = PropertySet(
+        {
+            "nticks": 5 * 365,
+            "beta": 1.5 / 7.0,  # 1.25 new infections per existing infection every 7 ticks
+            "r_recovery": 1.0 / 7.0,  # 7 days to recovery
+            "r_waning": 1.0 / 30.0,  # 30 days to susceptibility
+            **(params or {}),
+        }
+    )
     model = Model(scenario, p)
 
     betas = ValuesMap.from_scalar(p.beta, p.nticks, len(scenario))
@@ -92,6 +94,7 @@ def test_sirs() -> None:
 
 
 if __name__ == "__main__":
+
     def _parse_value(s: str) -> int | float:
         try:
             return int(s)